@@ -15,7 +15,6 @@ classifiers =
     Operating System :: OS Independent
     Programming Language :: Python :: 3
     Programming Language :: Python :: 3 :: Only
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
@@ -32,8 +31,8 @@ install_requires =
     click
     cloudpickle
     loky
-    pytask>=0.1.0
-python_requires = >=3.6
+    pytask>=0.1.7
+python_requires = >=3.7
 include_package_data = True
 package_dir = =src
 zip_safe = False
