@@ -4,7 +4,6 @@ from __future__ import annotations
 import sys
 import time
 from typing import Any
-from typing import Tuple
 
 import cloudpickle
 from _pytask.config import hookimpl
@@ -178,8 +177,8 @@ def _unserialize_and_execute_task(bytes_, show_locals, console_options):
 
 
 def _process_exception(
-    exc_info: Tuple[Any], show_locals: bool, console_options: ConsoleOptions
-) -> Tuple[Any]:
+    exc_info: tuple[Any], show_locals: bool, console_options: ConsoleOptions
+) -> tuple[Any]:
     exc_info = remove_internal_traceback_frames_from_exc_info(exc_info)
     traceback = Traceback.from_exception(*exc_info, show_locals=show_locals)
     segments = console.render(traceback, options=console_options)
